requirements.txt
pdfplumber
langchain
sentence-transformers
faiss-cpu
openai
streamlit
